Consolidate upsert_sample function
The `upsert_sample` function is implemented in three different places. This
consolidates those to a single standardized version of the function to reduce
redundancy. Also updates existing `upsert_sample` function calls to match the
new definition, as well as some minor code clean up.

This function is also being moved out of the `id3c.cli.command.etl` subpackage to
`id3c.db` in anticipation of also using it for a API development. Other "upsert"
functions are currently located in `etl/__init__.py` but are currently only used in
`etl` modules, whereas `db` contains more generic functions/classes used in `id3c.api`
and `id3c.cli`.
