+++ lib/id3c/cli/command/etl/__init__.py
@@ -384,42 +384,6 @@ def upsert_presence_absence(db: DatabaseSession,
     return presence_absence
 
 
-def upsert_sample(db: DatabaseSession,
-                  collection_identifier: str,
-                  encounter_id: int,
-                  details: dict) -> Any:
-    """
-    Upsert collected sample by its *collection_identifier*.
-
-    The provided *details* are merged (at the top-level only) into
-    the existing sample details, if any.
-    """
-    LOG.debug(f"Upserting sample collection «{collection_identifier}»")
-
-    data = {
-        "collection_identifier": collection_identifier,
-        "encounter_id": encounter_id,
-        "details": Json(details),
-    }
-
-    sample = db.fetch_row("""
-        insert into warehouse.sample (collection_identifier, encounter_id, details)
-            values (%(collection_identifier)s, %(encounter_id)s, %(details)s)
-
-        on conflict (collection_identifier) do update
-            set encounter_id = excluded.encounter_id,
-                details = coalesce(sample.details, '{}') || %(details)s
-
-        returning sample_id as id, identifier, collection_identifier, encounter_id
-        """, data)
-
-    assert sample.id, "Upsert affected no rows!"
-
-    LOG.info(f"Upserted sample {sample.id} with collection identifier «{sample.collection_identifier}»")
-
-    return sample
-
-
 def find_or_create_target(db: DatabaseSession, identifier: str, control: bool) -> Any:
     """
     Select presence_absence test target by *identifier*, or insert it if it doesn't exist.
