+++ lib/id3c/cli/command/etl/fhir.py
@@ -23,7 +23,7 @@ from fhir.resources.patient import Patient
 from fhir.resources.questionnaireresponse import QuestionnaireResponse
 from fhir.resources.specimen import Specimen
 from id3c.cli.command import with_database_session
-from id3c.db import find_identifier
+from id3c.db import find_identifier, upsert_sample
 from id3c.db.session import DatabaseSession
 from id3c.db.datatypes import Json
 from id3c.utils import getattrpath
@@ -671,92 +671,12 @@ def process_encounter_samples(db: DatabaseSession, encounter: Encounter, encount
         # XXX TODO: Improve details object here; the current approach produces
         # an object like {"coding": [{…}]} which isn't very useful.
         upsert_sample(db,
+            update_identifiers      = False,
             identifier              = sample_identifier,
             collection_identifier   = collection_identifier,
             collection_date         = collection_date,
             encounter_id            = encounter_id,
             additional_details      = additional_details)
-
-# Copied from etl manifest and edited to fit the needs for the FHIR etl.
-# We may want to consolidate `upsert_sample` at some point in the future,
-# but this was done to ensure that nothing goes wrong with the manifest etl.
-#       -Jover, 12 Feb 2020
-def upsert_sample(db: DatabaseSession,
-                  identifier: Optional[str],
-                  collection_identifier: Optional[str],
-                  collection_date: Optional[str],
-                  encounter_id: int,
-                  additional_details: dict) -> Any:
-    """
-    Upsert sample by its *identifier* and/or *collection_identifier*.
-
-    An existing sample has its *identifier*, *collection_identifier*,
-    and *encounter_id* updated if provided, and the provided
-    *additional_details* are merged (at the top-level only)
-    into the existing sample details, if any.
-
-    Raises an exception if there is more than one matching sample.
-    """
-    data = {
-        "identifier": identifier,
-        "collection_identifier": collection_identifier,
-        "collection_date": collection_date,
-        "encounter_id": encounter_id,
-        "additional_details": Json(additional_details),
-    }
-
-    # Look for existing sample(s)
-    with db.cursor() as cursor:
-        cursor.execute("""
-            select sample_id as id, identifier, collection_identifier, encounter_id
-              from warehouse.sample
-             where identifier = %(identifier)s
-                or collection_identifier = %(collection_identifier)s
-               for update
-            """, data)
-
-        samples = list(cursor)
-
-    # Nothing found → create
-    if not samples:
-        LOG.info("Creating new sample")
-        sample = db.fetch_row("""
-            insert into warehouse.sample (identifier, collection_identifier, collected, encounter_id, details)
-                values (%(identifier)s,
-                        %(collection_identifier)s,
-                        date_or_null(%(collection_date)s),
-                        %(encounter_id)s,
-                        %(additional_details)s)
-            returning sample_id as id, identifier, collection_identifier, encounter_id
-            """, data)
-
-    # One found → update
-    elif len(samples) == 1:
-        sample = samples[0]
-
-        LOG.info(f"Updating existing sample {sample.id}")
-        sample = db.fetch_row("""
-            update warehouse.sample
-               set identifier = coalesce(%(identifier)s, identifier),
-                   collection_identifier = coalesce(%(collection_identifier)s, collection_identifier),
-                   collected = coalesce(date_or_null(%(collection_date)s), collected),
-                   encounter_id = %(encounter_id)s,
-                   details = coalesce(details, '{}') || %(additional_details)s
-
-             where sample_id = %(sample_id)s
-
-            returning sample_id as id, identifier, collection_identifier, encounter_id
-            """,
-            { **data, "sample_id": sample.id })
-
-        assert sample.id, "Update affected no rows!"
-
-    # More than one found → error
-    else:
-        raise Exception(f"More than one sample matching sample and/or collection barcodes: {samples}")
-
-    return sample
-
 
 def encounter_age(encounter: Encounter, resources: Dict[str, List[DomainResource]]) -> Optional[str]:
     """
